Run firebase serve, switch to getDownloadURL, cleanup company edit page
In order for us to start using Firebase functions (which the importer will
use we need to start using Firebase serve). I set this up so npm start will
run both serve and start.

We were storing getDownloadURL but that URL contains a token that can be
revoked and changed. So instead I'm storing the path in firestore and then
calling getDownloadURL at runtime. I didn't reupload everyone's images yet
because I wanted to make sure everyone was okay with this change.

I cleaned up the admin company edit page since it wasn't using hooks for
everything but the AdminCompanyPage also lets you upload and looks nicer
so maybe we should move the edit page into the normal page?